Finish Code Changes hurray (i never usually commit like this)

+++ src/assignment2.cpp
@@ -1,6 +1,8 @@
 #include <iostream>
 #include <filesystem>
 #include <vector>
+#include <chrono>
+#include <thread>
 #include "commands.hpp"
 #include "history.hpp"
 #include "foldertools.hpp"
@@ -13,13 +15,13 @@ struct systemState
     std::vector<pid_t> runing_programs;
 };
 
-void executeCommand(HistoryEntry e, systemState &current_state);
+void executeCommand(HistoryEntry &e, systemState &current_state);
+
 int main()
 {
     systemState current_state;
     current_state.current_path = std::filesystem::current_path();
     current_state.history = getHistoryFromFile("history.txt");
-    // TODO: read history
     Command lastCommand;
 #ifdef DEBUG
     std::cout << "cwd: " << current_state.current_path.generic_string() << std::endl;
@@ -34,18 +36,18 @@ int main()
         HistoryEntry e;
         e.args = args;
         e.command = command;
-        if (command != Command::NONE && command != Command::REPLAY)
+        if (command != Command::NONE)
         {
             current_state.history.push_back(e);
         }
         executeCommand(e, current_state);
-        lastCommand = command;
+        lastCommand = e.command;
     } while (lastCommand != Command::BYEBYE);
 
     dumpHistoryToFile("history.txt", current_state.history);
 }
 
-void executeCommand(HistoryEntry e, systemState &current_state)
+void executeCommand(HistoryEntry &e, systemState &current_state)
 {
     Command command = e.command;
     std::string args = e.args;
@@ -55,6 +57,11 @@ void executeCommand(HistoryEntry e, systemState &current_state)
     case Command::MOVETODIR:
     {
         bool success = moveToFolder(args, current_state.current_path);
+        if (!success)
+        {
+            std::cout << "Could not move to that directory." << std::endl;
+            break;
+        }
         break;
     }
     case Command::WHEREAMI:
@@ -73,21 +80,29 @@ void executeCommand(HistoryEntry e, systemState &current_state)
     case Command::REPLAY:
     {
         int idx = std::stoi(args);
-        HistoryEntry previous = current_state.history[current_state.history.size() - idx - 1];
+        HistoryEntry previous = current_state.history[current_state.history.size() - idx];
+        
+        std::cout << commandToString(previous.command) << previous.args << std::endl;
 
+        if (previous.command == Command::BYEBYE)
+        {
+            e.command = previous.command;
+            return;
+        }
         executeCommand(previous, current_state);
         break;
     }
     case Command::START:
-        startProgramAndWait(args);
+        startProgramAndWait(args, current_state.current_path);
         break;
     case Command::BACKGROUND:
     {
-        pid_t new_program = startProgram(args);
+        pid_t new_program = startProgram(args, current_state.current_path);
         if(new_program != -1)
         {
-            std::cout << "started with pid: " << std::to_string(new_program) << std::endl;
+            std::cout << "Started with pid: " << std::to_string(new_program) << std::endl;
             current_state.runing_programs.push_back(new_program);
+            std::this_thread::sleep_for(std::chrono::milliseconds(500));
         }
         break;
     }
@@ -115,30 +130,54 @@ void executeCommand(HistoryEntry e, systemState &current_state)
 
         for (int i = 0; i < numTimes; i++)
         {
-            pid_t pid = startProgram(command_string);
-            spawned.push_back(pid);
+            pid_t pid = startProgram(command_string, current_state.current_path);
+            if (pid != -1)
+            {
+                spawned.push_back(pid);
+            }
         }
+        std::this_thread::sleep_for(std::chrono::milliseconds(500));
 
         std::cout << "PIDs: ";
         for (int i = 0; i < spawned.size() - 1; i++)
         {
             std::cout << std::to_string(spawned[i]) << ", ";
+            current_state.runing_programs.push_back(spawned[i]);
         }
         std::cout << std::to_string(spawned[spawned.size() - 1]) << std::endl;
+        current_state.runing_programs.push_back(spawned[spawned.size() - 1]);
 
         break;
     }
     case Command::DALEKALL:
     {
+        if (current_state.runing_programs.size() == 0)
+        {
+            std::cout << "There are no running background processes" << std::endl;
+            break;
+        }
+        
+        std::cout << "Exterminating " << std::to_string(current_state.runing_programs.size()) << (current_state.runing_programs.size() == 1 ? " process: " : " processes: ");
+
         for (int i = 0; i < current_state.runing_programs.size(); i++)
         {
             pid_t pid = current_state.runing_programs[i];
+            std::cout << std::to_string(pid);
+            if (i != current_state.runing_programs.size() - 1)
+            {
+                std::cout << " ";
+            }
             killProcess(pid);
         }
+        std::cout << std::endl;
         current_state.runing_programs.clear();
         break;
     }
     case Command::BYEBYE:
+    // {
+    //     dumpHistoryToFile("history.txt", current_state.history);
+    //     exit(0);
+    // }
         break;
     default:
         std::cout << "Command unknown" << std::endl;
